Create Controllers/AbsenceCheckController.cs: 1. For Accept or Reject Absence Table. 2. Only can read date which checkID=1.
Create Controllers/AbsenceMISController.cs:
1.For MIS to alter Absence Table data.
2.For MIS to read data weather data Dispose is true or false.

Alter Controllers/AbsenceTableController.cs:
1. For the way in how absenceTable get data.
2. Create a method for judge EmployeeOccu.

+++ BookStore(Web)_LC/BookStore(Web)/Controllers/AbsenceTableController.cs
@@ -15,13 +15,15 @@ namespace BookStore_Web_.Controllers
         BookStoreObject_Entities db = new BookStoreObject_Entities();
         Repository<Absence_Table> absenceTableRepository = null;
         Repository<Absence_Type> absenceTypeRepository = null;
-        int EmpID = HomeController.EmpID = 3;
-        //int? page, Dept, Occu;
-        //string sortBy;
+        IEnumerable<BookStore_Web_.Models.Absence_Table> absenceTable = null;
+        int EmpID = HomeController.EmpID = 9;
+        int EmpOccu;
         public AbsenceTableController()
         {
             absenceTableRepository = new Repository<Absence_Table>(new BookStoreObject_Entities());
             absenceTypeRepository = new Repository<Absence_Type>(new BookStoreObject_Entities());
+            absenceTable = absenceTableRepository.GetAll().Where(p => p.Dispose == false);
+            EmpOccu = db.Emp_Information.Where(p => p.EmployeeID == EmpID).Select(p => p.Emp_OccupationID).FirstOrDefault();
         }
         
         public ActionResult Index(int? page)
@@ -34,35 +36,16 @@ namespace BookStore_Web_.Controllers
             ViewBag.SortStartDateDesc = "StartDatedesc";
             ViewBag.SortStatus = "Status";
             ViewBag.SortStatusDesc = "Statusdesc";
-            var absenceTable = absenceTableRepository.GetAll();
             //判斷權限
-            var q = db.Emp_Information.Where(p => p.EmployeeID == EmpID).Select(p => p.Emp_OccupationID).FirstOrDefault();
-            if (q == 3)
-                return View(absenceTable.Where(p => p.EmployeeID == EmpID).ToList().ToPagedList(page ?? 1, 5));
-            else
-                return View(absenceTable.ToList().ToPagedList(page ?? 1, 5));
+            absenceTable = EmployeeOccu(EmpOccu);
+            return View(absenceTable.ToPagedList(page ?? 1, 5));
         }
 
-        public ActionResult IndexPager(int? page, int? Dept=0, int Occu=0)
+        public ActionResult IndexPager(int? page, int Dept=0, int Occu=0)
         {            
-            var absenceTable = absenceTableRepository.GetAll();
             //判斷權限
-            var q = db.Emp_Information.Where(p => p.EmployeeID == EmpID).Select(p => p.Emp_OccupationID).FirstOrDefault();
-            if (q == 3)
-            {
-                absenceTable = absenceTable.Where(p => p.EmployeeID == EmpID).ToList();
-            }
-            else
-            {
-                if (Dept != 0 && Occu != 0)               
-                    absenceTable = absenceTable.Where(p => p.Emp_Information.Emp_DeptID == Dept && p.Emp_Information.Emp_OccupationID == Occu).ToList();                
-                else if (Dept != 0 && Occu == 0)                    
-                    absenceTable = absenceTable.Where(p => p.Emp_Information.Emp_DeptID == Dept).ToList();                
-                else if (Occu != 0 && Dept == 0)                
-                    absenceTable = absenceTable.Where(p => p.Emp_Information.Emp_OccupationID == Occu).ToList();                
-                else                
-                    absenceTable = absenceTable.ToList();  
-            }
+            absenceTable = EmployeeOccu(EmpOccu, Dept, Occu);
+
             page = SetPage(page, absenceTable.Count());
             return PartialView(absenceTable.ToPagedList(page ?? 1, 5));
         }
@@ -79,11 +62,36 @@ namespace BookStore_Web_.Controllers
                 return page;
         }
 
+        IEnumerable<Absence_Table> EmployeeOccu(int empOccu, int Dept=0, int Occu=0) {
+            
+            if (empOccu == 3)//一般員工
+                absenceTable = absenceTable.Where(p => p.EmployeeID == EmpID).ToList();
+            else if (empOccu == 2)//高權限
+            {
+                absenceTable = absenceTable.Where(p => p.Emp_Information.Emp_Report == EmpID);//只會看到 "report"是自己EmpID的請假單                       
+            }
+            else if (empOccu == 1)//總經理
+            {
+                absenceTable = absenceTable.Where(p => p.Emp_Information.Emp_Report == EmpID);//只會看到 "report"是自己EmpID的請假單
+            }
+            
+            if (Dept != 0 && Occu != 0)
+                absenceTable = absenceTable.Where(p => p.Emp_Information.Emp_DeptID == Dept && p.Emp_Information.Emp_OccupationID == Occu).ToList();
+            else if (Dept != 0 && Occu == 0)
+                absenceTable = absenceTable.Where(p => p.Emp_Information.Emp_DeptID == Dept).ToList();
+            else if (Occu != 0 && Dept == 0)
+                absenceTable = absenceTable.Where(p => p.Emp_Information.Emp_OccupationID == Occu).ToList();
+            else
+                absenceTable = absenceTable.ToList();
+
+
+            return absenceTable;
+        }
 
         //[ChildActionOnly]
         public ActionResult IndexPartial(int? page, string sortBy, int Dept=0, int Occu=0)
         {
-            var absenceTable = absenceTableRepository.GetAll();
+            #region for sortBy
             switch (sortBy)
             {
                 case "AbsenceTableNo":
@@ -114,107 +122,69 @@ namespace BookStore_Web_.Controllers
                     absenceTable = absenceTable.OrderBy(p => p.Absence_No);
                     break;
             }
-            //if (Request.UrlReferrer.Query != "" && Request.UrlReferrer.Query.Split('=')[1] != "0")
-            //{
-            //    page = Convert.ToInt32(Request.UrlReferrer.Query.Split('=')[1]);
-            //}
-            //var q = db.Emp_Information.Where(p => p.EmployeeID == EmpID).Select(p => p.Emp_OccupationID).FirstOrDefault();
-            //if (q == 3)
-            //    return PartialView(absenceTable.Where(p => p.EmployeeID == EmpID).ToList().ToPagedList(page ?? 1, 5));
-            //else
-            //{
-            //    if (Dept != null && Occu != null)
-            //    {
-            //        if (Dept != 0 && Occu != 0)
-            //            return PartialView(absenceTable.Where(p => p.Emp_Information.Emp_DeptID == Dept && p.Emp_Information.Emp_OccupationID == Occu).ToList().ToPagedList(page ?? 1, 5));
-            //        else if (Dept != 0 && Occu == 0)
-            //        {
-            //            return PartialView(absenceTable.Where(p => p.Emp_Information.Emp_DeptID == Dept).ToList().ToPagedList(page ?? 1, 5));
-            //        }
-            //        else if (Occu != 0 && Dept == 0)
-            //        {
-            //            return PartialView(absenceTable.Where(p => p.Emp_Information.Emp_OccupationID == Occu).ToList().ToPagedList(page ?? 1, 5));
-            //        }
-            //        else
-            //            return PartialView(absenceTable.ToList().ToPagedList(page ?? 1, 5));
-            //    }
-            //    else
-            //        absenceTable = absenceTable.ToList();
-            //    page = SetPage(page, absenceTable.Count());
-            //        return PartialView(absenceTable.ToPagedList(page ?? 1, 5));
-            //}
-            var q = db.Emp_Information.Where(p => p.EmployeeID == EmpID).Select(p => p.Emp_OccupationID).FirstOrDefault();
-            if (q == 3)
-            {
-                absenceTable = absenceTable.Where(p => p.EmployeeID == EmpID).ToList();
-            }
-            else
-            {
-                if (Dept != 0 && Occu != 0)
-                    absenceTable = absenceTable.Where(p => p.Emp_Information.Emp_DeptID == Dept && p.Emp_Information.Emp_OccupationID == Occu).ToList();
-                else if (Dept != 0 && Occu == 0)
-                    absenceTable = absenceTable.Where(p => p.Emp_Information.Emp_DeptID == Dept).ToList();
-                else if (Occu != 0 && Dept == 0)
-                    absenceTable = absenceTable.Where(p => p.Emp_Information.Emp_OccupationID == Occu).ToList();
-                else
-                    absenceTable = absenceTable.ToList();
-            }
+            #endregion
+
+            //判斷權限
+            absenceTable = EmployeeOccu(EmpOccu, Dept, Occu);
+
             page = SetPage(page, absenceTable.Count());
             return PartialView(absenceTable.ToPagedList(page ?? 1, 5));
         }
 
         public ActionResult Dept()
         {
-            var q = db.Dept_Information.Select(p => new
+            var dept = db.Dept_Information.Select(p => new
             {
                 DeptID = p.Emp_DeptID,
                 DeptName = p.Dept_Name
             });
-            return Json(q.ToList(),JsonRequestBehavior.AllowGet);
+            return Json(dept.ToList(), JsonRequestBehavior.AllowGet);
 
         }
 
         public ActionResult Occu()
         {
-            var q = db.Emp_Occupation.Select(p => new
+            var occu = db.Emp_Occupation.Select(p => new
             {
                 OccuID = p.Emp_OccupationID,
                 OccuName = p.Emp_OccupationName
             });
-            return Json(q.ToList(), JsonRequestBehavior.AllowGet);
+            return Json(occu.ToList(), JsonRequestBehavior.AllowGet);
 
         }
 
-        public ActionResult Add(Absence_Table absenceTable)
+        public ActionResult Add(Absence_Table _absenceTable)
         {
             ViewBag.AbsenceType = new SelectList(absenceTypeRepository.GetAll().ToList(), "Absence_ID", "Absence_Type1");
-            //ViewBag.StartDate = DateTime.Now.ToString("yyyy/MM/dd tt mm:ss");
             if (ModelState.IsValid)
             {
-                absenceTable.EmployeeID = EmpID;
-                absenceTable.Check_ID = 1;
-                absenceTable.Dispose = false;
-                absenceTable.StartDate = Convert.ToDateTime(string.Format("{0} {1}", absenceTable.StartDate.ToShortDateString(), Request.Form["StartDatetime"]));
-                absenceTable.EndDate = Convert.ToDateTime(string.Format("{0} {1}", absenceTable.EndDate.ToShortDateString(), Request.Form["EndDatetime"]));
+                _absenceTable.EmployeeID = EmpID;
+                _absenceTable.Check_ID = 1;
+                _absenceTable.Dispose = false;
+                _absenceTable.StartDate = Convert.ToDateTime(string.Format("{0} {1}", _absenceTable.StartDate.ToShortDateString(), Request.Form["StartDatetime"]));
+                _absenceTable.EndDate = Convert.ToDateTime(string.Format("{0} {1}", _absenceTable.EndDate.ToShortDateString(), Request.Form["EndDatetime"]));
                 HttpPostedFileBase file = Request.Files["Certificate_Doc"];
                 if (file != null && file.ContentLength > 0)
                 {
                     var imgSize = file.ContentLength;
                     byte[] imgByte = new Byte[imgSize];
                     file.InputStream.Read(imgByte, 0, imgSize);
-                    absenceTable.Certificate_Doc = imgByte;
+                    _absenceTable.Certificate_Doc = imgByte;
                 }
-                absenceTableRepository.Create(absenceTable);
+
+                absenceTableRepository.Create(_absenceTable);
                 return RedirectToAction("Index");
             }
             
             return View();
         }
 
-        public ActionResult Delete(int id)            
+        public ActionResult Delete(int id)
         {
             var delete = absenceTableRepository.GetByID(id);
-            absenceTableRepository.Delete(delete);
+            if (delete.Check_ID != 2)
+                absenceTableRepository.Delete(delete); ;
+
             return RedirectToAction("Index");
         }
 
@@ -228,13 +198,13 @@ namespace BookStore_Web_.Controllers
         }
 
         [HttpPost]
-        public ActionResult Edit(Absence_Table Absent)
+        public ActionResult Edit(Absence_Table _absenceTable)
         {            
-            Absence_Table ab = absenceTableRepository.GetByID(Absent.Absence_No);
-            ab.Absence_ID = Absent.Absence_ID;
-            ab.StartDate = Absent.StartDate;
-            ab.EndDate = Absent.EndDate;
-            ab.Reason = Absent.Reason;
+            Absence_Table ab = absenceTableRepository.GetByID(_absenceTable.Absence_No);
+            ab.Absence_ID = _absenceTable.Absence_ID;
+            ab.StartDate = _absenceTable.StartDate;
+            ab.EndDate = _absenceTable.EndDate;
+            ab.Reason = _absenceTable.Reason;
             HttpPostedFileBase file = Request.Files["CertificateDoc"];
             if(file.ContentLength>0)
             {
@@ -257,7 +227,6 @@ namespace BookStore_Web_.Controllers
                 return new FileContentResult(img, "image/jpeg");
            
         }
-
 
     }
 }
